chore(amplify): add amplify configuration
