+++ next.config.js
@@ -4,6 +4,9 @@ module.exports = withCss({
   serverRuntimeConfig: { // Only available on server side
   },
   publicRuntimeConfig: { // Available on both server and client
+    AWS_REGION: process.env.AWS_REGION,
+    USER_POOL_ID: process.env.USER_POOL_ID,
+    USER_POOL_CLIENT_ID: process.env.USER_POOL_CLIENT_ID,
   },
   webpack (config) {
     config.module.rules.push({
