+++ package.json
@@ -37,6 +37,7 @@
   "dependencies": {
     "@zeit/next-css": "^1.0.1",
     "apollo-boost": "^0.1.16",
+    "aws-amplify": "^1.1.10",
     "cookie-parser": "^1.4.3",
     "express": "^4.16.4",
     "formik": "^1.3.1",
