+++ utils/auth.js
@@ -1,21 +1,26 @@
-import { medic, beneficiary, nurse, error, admin } from '../components/Layout/rolesProperties.json'
+import Amplify from 'aws-amplify'
+import getConfig from 'next/config'
 
-const getPropertiesFromRole = (userRole) => {
-  if ( userRole === 'medic' ) {
-    return medic
-  }
-  else if ( userRole === 'beneficiary' ) {
-    return beneficiary
-  }
-  else if ( userRole === 'nurse' ) {
-    return nurse
-  }
-  else if ( userRole === 'admin' ) {
-    return admin
-  }
-  else {
-    return error
-  }
-}
+const config = getConfig()
+const {
+  AWS_REGION,
+  USER_POOL_ID,
+  USER_POOL_CLIENT_ID,
+} = (config) ? config.publicRuntimeConfig : {}
 
-export { getPropertiesFromRole }
+const amplifyConfigure = Amplify.configure({
+  Auth: {
+    region: AWS_REGION,
+    userPoolId: USER_POOL_ID,
+    userPoolWebClientId: USER_POOL_CLIENT_ID,
+    cookieStorage: {
+      domain: '.michellesagnelli.com',
+      path: '/',
+      expires: 365,
+      secure: true,
+    },
+    authenticationFlowType: 'USER_PASSWORD_AUTH',
+  },
+})
+
+export { amplifyConfigure }
